+++ packages/entities/entities-consumer-groups/src/components/ConsumerGroupForm.vue
@@ -4,7 +4,6 @@
       :can-submit="isFormValid && changesExist"
       :config="config"
       :edit-id="consumerGroupId"
-      :enable-terraform="enableTerraform"
       :entity-type="SupportedEntityType.ConsumerGroup"
       :error-message="state.errorMessage || fetchError || preValidateErrorMessage"
       :fetch-url="fetchUrl"
@@ -126,14 +125,6 @@ const props = defineProps({
     type: String,
     required: false,
     default: '',
-  },
-  /**
-   * Enable display of Terraform code
-   * Guarded by FF: khcp-12445-terraform-config-details
-   */
-  enableTerraform: {
-    type: Boolean,
-    default: false,
   },
 })