+++ packages/entities/entities-gateway-services/src/components/GatewayServiceForm.vue
@@ -4,7 +4,6 @@
       :can-submit="canSubmit"
       :config="config"
       :edit-id="gatewayServiceId"
-      :enable-terraform="enableTerraform"
       :entity-type="SupportedEntityType.GatewayService"
       :error-message="form.errorMessage"
       :fetch-url="fetchUrl"
@@ -429,14 +428,6 @@ const props = defineProps({
   hideSectionsInfo: {
     type: Boolean,
     required: false,
-    default: false,
-  },
-  /**
-   * Enable display of Terraform code
-   * Guarded by FF: khcp-12445-terraform-config-details
-   */
-  enableTerraform: {
-    type: Boolean,
     default: false,
   },
 })