+++ packages/entities/entities-key-sets/src/components/KeySetForm.vue
@@ -4,7 +4,6 @@
       :can-submit="canSubmit"
       :config="config"
       :edit-id="keySetId"
-      :enable-terraform="enableTerraform"
       :entity-type="SupportedEntityType.KeySet"
       :error-message="form.errorMessage"
       :fetch-url="fetchUrl"
@@ -100,14 +99,6 @@ const props = defineProps({
     type: String,
     required: false,
     default: '',
-  },
-  /**
-   * Enable display of Terraform code
-   * Guarded by FF: khcp-12445-terraform-config-details
-   */
-  enableTerraform: {
-    type: Boolean,
-    default: false,
   },
 })