+++ packages/entities/entities-keys/src/components/KeyConfigCard.vue
@@ -4,7 +4,6 @@
       :config="config"
       :config-card-doc="configCardDoc"
       :config-schema="configSchema"
-      :enable-terraform="enableTerraform"
       :entity-type="SupportedEntityType.Key"
       :fetch-url="fetchUrl"
       :hide-title="hideTitle"
@@ -130,14 +129,6 @@ const props = defineProps({
   keySetId: {
     type: String as PropType<string | null>,
     default: null,
-  },
-  /**
-   * Enable display of Terraform code
-   * Guarded by FF: khcp-12445-terraform-config-details
-   */
-  enableTerraform: {
-    type: Boolean,
-    default: false,
   },
 })