+++ packages/entities/entities-keys/src/components/KeyForm.vue
@@ -4,7 +4,6 @@
       :can-submit="canSubmit"
       :config="config"
       :edit-id="keyId"
-      :enable-terraform="enableTerraform"
       :entity-type="SupportedEntityType.Key"
       :error-message="form.errorMessage || fetchKeySetsErrorMessage"
       :fetch-url="fetchUrl"
@@ -229,14 +228,6 @@ const props = defineProps({
     type: String,
     required: false,
     default: '',
-  },
-  /**
-   * Enable display of Terraform code
-   * Guarded by FF: khcp-12445-terraform-config-details
-   */
-  enableTerraform: {
-    type: Boolean,
-    default: false,
   },
 })