+++ packages/entities/entities-snis/src/components/SniForm.vue
@@ -4,7 +4,6 @@
       :can-submit="canSubmit"
       :config="config"
       :edit-id="sniId"
-      :enable-terraform="enableTerraform"
       :entity-type="SupportedEntityType.SNI"
       :error-message="form.errorMessage || fetchCertsErrorMessage"
       :fetch-url="fetchUrl"
@@ -149,14 +148,6 @@ const props = defineProps({
     type: String,
     required: false,
     default: '',
-  },
-  /**
-   * Enable display of Terraform code
-   * Guarded by FF: khcp-12445-terraform-config-details
-   */
-  enableTerraform: {
-    type: Boolean,
-    default: false,
   },
 })