+++ packages/entities/entities-upstreams-targets/src/components/TargetsList.vue
@@ -2,13 +2,13 @@
   <div class="kong-ui-entities-targets-list">
     <EntityBaseTable
       :cache-identifier="cacheIdentifier"
-      disable-pagination-page-jump
       :disable-sorting="disableSorting"
       :empty-state-options="emptyStateOptions"
       enable-entity-actions
       :error-message="errorMessage"
       :fetcher="fetcher"
       :fetcher-cache-key="fetcherCacheKey"
+      :hide-toolbar="hideTableToolbar"
       pagination-type="offset"
       preferences-storage-key="kong-ui-entities-targets-list"
       :table-headers="tableHeaders"
@@ -273,6 +273,8 @@ const resetPagination = (): void => {
  * loading, Error, Empty state
  */
 const errorMessage = ref<TableErrorMessage>(null)
+// hide table toolbar in initial loading state or when no records are found
+const hideTableToolbar = ref<boolean>(false)
 
 /**
  * Create target
@@ -411,6 +413,12 @@ watch(fetcherState, (state) => {
   // reset `hasData` to show/hide the teleported Create button
   if (Array.isArray(state?.response?.data)) {
     hasData.value = state.response!.data.length > 0
+  }
+
+  if (state.status === FetcherStatus.NoRecords) {
+    hideTableToolbar.value = true
+  } else {
+    hideTableToolbar.value = false
   }
 
   if (state.status === FetcherStatus.Error) {