+++ packages/entities/entities-vaults/src/components/VaultForm.vue
@@ -4,7 +4,6 @@
       :can-submit="isFormValid && changesExist"
       :config="config"
       :edit-id="vaultId"
-      :enable-terraform="enableTerraform"
       :entity-type="SupportedEntityType.Vault"
       :error-message="form.errorMessage"
       :fetch-url="fetchUrl"
@@ -571,14 +570,6 @@ const props = defineProps({
     type: String,
     required: false,
     default: '',
-  },
-  /**
-   * Enable display of Terraform code
-   * Guarded by FF: khcp-12445-terraform-config-details
-   */
-  enableTerraform: {
-    type: Boolean,
-    default: false,
   },
 })