+++ packages/analytics/analytics-metric-provider/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.4.37](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@8.4.36...@kong-ui-public/analytics-metric-provider@8.4.37) (2024-10-17)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
+## [8.4.36](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@8.4.35...@kong-ui-public/analytics-metric-provider@8.4.36) (2024-10-16)
+
+
+### Bug Fixes
+
+* ktabledata adoption [KHCP-13280] ([#1660](https://github.com/Kong/public-ui-components/issues/1660)) ([31ca10e](https://github.com/Kong/public-ui-components/commit/31ca10eb93f51bf75bbe4419f2ddda0fdf47279e))
+
+
+
+
+
 ## [8.4.35](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@8.4.34...@kong-ui-public/analytics-metric-provider@8.4.35) (2024-10-14)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-metric-provider