+++ packages/core/app-layout/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.49](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/app-layout@4.2.48...@kong-ui-public/app-layout@4.2.49) (2024-10-17)
+
+**Note:** Version bump only for package @kong-ui-public/app-layout
+
+
+
+
+
+## [4.2.48](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/app-layout@4.2.47...@kong-ui-public/app-layout@4.2.48) (2024-10-16)
+
+
+### Bug Fixes
+
+* ktabledata adoption [KHCP-13280] ([#1660](https://github.com/Kong/public-ui-components/issues/1660)) ([31ca10e](https://github.com/Kong/public-ui-components/commit/31ca10eb93f51bf75bbe4419f2ddda0fdf47279e))
+
+
+
+
+
 ## [4.2.47](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/app-layout@4.2.46...@kong-ui-public/app-layout@4.2.47) (2024-10-14)
 
 **Note:** Version bump only for package @kong-ui-public/app-layout