+++ packages/core/documentation/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.73](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/documentation@0.7.72...@kong-ui-public/documentation@0.7.73) (2024-10-17)
+
+**Note:** Version bump only for package @kong-ui-public/documentation
+
+
+
+
+
+## [0.7.72](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/documentation@0.7.71...@kong-ui-public/documentation@0.7.72) (2024-10-17)
+
+**Note:** Version bump only for package @kong-ui-public/documentation
+
+
+
+
+
+## [0.7.71](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/documentation@0.7.70...@kong-ui-public/documentation@0.7.71) (2024-10-16)
+
+
+### Bug Fixes
+
+* ktabledata adoption [KHCP-13280] ([#1660](https://github.com/Kong/public-ui-components/issues/1660)) ([31ca10e](https://github.com/Kong/public-ui-components/commit/31ca10eb93f51bf75bbe4419f2ddda0fdf47279e))
+
+
+
+
+
 ## [0.7.70](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/documentation@0.7.69...@kong-ui-public/documentation@0.7.70) (2024-10-14)
 
 **Note:** Version bump only for package @kong-ui-public/documentation