+++ packages/entities/entities-certificates/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.6.35](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-certificates@3.6.34...@kong-ui-public/entities-certificates@3.6.35) (2024-10-17)
+
+**Note:** Version bump only for package @kong-ui-public/entities-certificates
+
+
+
+
+
+## [3.6.34](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-certificates@3.6.33...@kong-ui-public/entities-certificates@3.6.34) (2024-10-17)
+
+**Note:** Version bump only for package @kong-ui-public/entities-certificates
+
+
+
+
+
+## [3.6.33](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-certificates@3.6.32...@kong-ui-public/entities-certificates@3.6.33) (2024-10-16)
+
+
+### Bug Fixes
+
+* ktabledata adoption [KHCP-13280] ([#1660](https://github.com/Kong/public-ui-components/issues/1660)) ([31ca10e](https://github.com/Kong/public-ui-components/commit/31ca10eb93f51bf75bbe4419f2ddda0fdf47279e))
+
+
+
+
+
 ## [3.6.32](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-certificates@3.6.31...@kong-ui-public/entities-certificates@3.6.32) (2024-10-14)
 
 **Note:** Version bump only for package @kong-ui-public/entities-certificates