+++ packages/entities/entities-gateway-services/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.8.20](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@3.8.19...@kong-ui-public/entities-gateway-services@3.8.20) (2024-10-17)
+
+**Note:** Version bump only for package @kong-ui-public/entities-gateway-services
+
+
+
+
+
+## [3.8.19](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@3.8.18...@kong-ui-public/entities-gateway-services@3.8.19) (2024-10-17)
+
+**Note:** Version bump only for package @kong-ui-public/entities-gateway-services
+
+
+
+
+
+## [3.8.18](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@3.8.17...@kong-ui-public/entities-gateway-services@3.8.18) (2024-10-16)
+
+
+### Bug Fixes
+
+* ktabledata adoption [KHCP-13280] ([#1660](https://github.com/Kong/public-ui-components/issues/1660)) ([31ca10e](https://github.com/Kong/public-ui-components/commit/31ca10eb93f51bf75bbe4419f2ddda0fdf47279e))
+
+
+
+
+
 ## [3.8.17](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@3.8.16...@kong-ui-public/entities-gateway-services@3.8.17) (2024-10-14)
 
 **Note:** Version bump only for package @kong-ui-public/entities-gateway-services