+++ packages/entities/entities-vaults/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.13.27](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-vaults@3.13.26...@kong-ui-public/entities-vaults@3.13.27) (2024-10-17)
+
+**Note:** Version bump only for package @kong-ui-public/entities-vaults
+
+
+
+
+
+## [3.13.26](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-vaults@3.13.25...@kong-ui-public/entities-vaults@3.13.26) (2024-10-17)
+
+**Note:** Version bump only for package @kong-ui-public/entities-vaults
+
+
+
+
+
+## [3.13.25](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-vaults@3.13.24...@kong-ui-public/entities-vaults@3.13.25) (2024-10-16)
+
+
+### Bug Fixes
+
+* ktabledata adoption [KHCP-13280] ([#1660](https://github.com/Kong/public-ui-components/issues/1660)) ([31ca10e](https://github.com/Kong/public-ui-components/commit/31ca10eb93f51bf75bbe4419f2ddda0fdf47279e))
+
+
+
+
+
 ## [3.13.24](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-vaults@3.13.23...@kong-ui-public/entities-vaults@3.13.24) (2024-10-14)
 
 **Note:** Version bump only for package @kong-ui-public/entities-vaults