+++ packages/portal/document-viewer/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.40](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/document-viewer@2.1.39...@kong-ui-public/document-viewer@2.1.40) (2024-10-17)
+
+**Note:** Version bump only for package @kong-ui-public/document-viewer
+
+
+
+
+
+## [2.1.39](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/document-viewer@2.1.38...@kong-ui-public/document-viewer@2.1.39) (2024-10-16)
+
+
+### Bug Fixes
+
+* ktabledata adoption [KHCP-13280] ([#1660](https://github.com/Kong/public-ui-components/issues/1660)) ([31ca10e](https://github.com/Kong/public-ui-components/commit/31ca10eb93f51bf75bbe4419f2ddda0fdf47279e))
+
+
+
+
+
 ## [2.1.38](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/document-viewer@2.1.37...@kong-ui-public/document-viewer@2.1.38) (2024-10-14)
 
 **Note:** Version bump only for package @kong-ui-public/document-viewer