+++ packages/portal/spec-renderer/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/spec-renderer@2.3.1...@kong-ui-public/spec-renderer@2.3.2) (2024-10-17)
+
+**Note:** Version bump only for package @kong-ui-public/spec-renderer
+
+
+
+
+
+## [2.3.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/spec-renderer@2.3.0...@kong-ui-public/spec-renderer@2.3.1) (2024-10-16)
+
+
+### Bug Fixes
+
+* ktabledata adoption [KHCP-13280] ([#1660](https://github.com/Kong/public-ui-components/issues/1660)) ([31ca10e](https://github.com/Kong/public-ui-components/commit/31ca10eb93f51bf75bbe4419f2ddda0fdf47279e))
+
+
+
+
+
 # [2.3.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/spec-renderer@2.2.51...@kong-ui-public/spec-renderer@2.3.0) (2024-10-15)